\name{data_GEUVADIS_combine}
\alias{data_GEUVADIS_combine}
\docType{data}
\title{
GEUVADIS data with combine 62 eQTLs-genes
}
\description{
GEUVADIS (Lappalainen et al., 2013) data combine with slelected 62 eQTLs-genes sets from Badsha and Fu, 2019.
}

\details{
The data set contains 373 samples and 194 eQTLs-gene sets (62 eQTLs and 132 genes).
}

\value{
Matrix
}

\references{
1. Lappalainen, T. et al. Transcriptome and genome sequencing uncovers functional variation in humans. Nature 501, 506-511 (2013).

2. Md Bahadur Badsha and Audrey Qiuyan Fu. Learning causal biological networks with the principle of Mendelian randomization. Frontiers in Genetics 2019, 10(460).
}

