\name{Data_withoutlier}
\alias{Data_withoutlier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Example Data with outlier
}
\description{
This is the data with outlier
}

\value{
Matrix
}

\author{
Md Bahadur Badsha <mdbadsha@uidaho.edu>,Evan A Martin <mart9986@vandals.uidaho.edu> and Audrey Qiuyan Fu <audreyf@uidaho.edu>
}

\examples{
Data_withoutlier
}
