\name{Data_withoutoutlier}
\alias{Data_withoutoutlier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Example Data without outlier
}
\description{
This is the data without outlier
}

\value{
Matrix
}

\author{
Md. Bahadur Badsha (mdbadsha@uidaho.edu)
}

\examples{
Data_withoutoutlier
}
