\name{simu_data_layered}
\alias{simu_data_layered}
\title{
Data for the layered model
}

\description{
Data simulated under the layered Model.
}

\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\seealso{
\link{SimulateData}.
}
