\name{simu_data_M0}
\alias{simu_data_M0}

\title{
Data for Model 0
}
\description{
Data simulated under Model 0.
}

\details{
The columns of the data matrix are the genetic variant (V node) and phenotype nodes (T nodes).
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\seealso{
\link{SimulateData}.
}

