\name{simu_data_M2}
\alias{simu_data_M2}
\title{
Data for Model 2
}
\description{
Data simulated under Model 2.
}

\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}

\value{
Matrix
}
\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}
\seealso{
\link{SimulateData}.
}
