\name{simu_data_M3}
\alias{simu_data_M3}
\title{
Data for Model 3
}
\description{
Data simulated under Model 3.
}
\details{
The columns of the data matrix are the genetic variant (V node) and phenotype nodes (T nodes).
}
\value{
Matrix
}

\author{
Md Bahadur Badsha (mbbadshar@gmail.com)
}
\seealso{
\link{SimulateData}.
}
