\name{simu_data_M4}
\alias{simu_data_M4}
\title{
Data for Model 4
}
\description{
Data simulated under Model 4.
}
\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}
\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}
\seealso{
\link{SimulateData}.
}

