\name{Data_withoutoutlier}
\alias{Data_withoutoutlier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Example Data without outliers
}
\description{
This is the data without outliers.
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\examples{
Data_withoutoutlier
}
