\name{Data_withoutoutlier}
\alias{Data_withoutoutlier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Example Data without outliers
}
\description{
This is the data without outliers.
}

\value{
Matrix
}

\author{
Md Bahadur Badsha <mdbadsha@uidaho.edu>, Evan A Martin <mart9986@vandals.uidaho.edu>, and Audrey Qiuyan Fu <audreyf@uidaho.edu>
}

\examples{
Data_withoutoutlier
}
