\name{simu_data_M0}
\alias{simu_data_M0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Data for Model 0
}
\description{
Data simulated under Model 0.
}

\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\seealso{
\link{SimulateData}.
}

