\name{simu_data_M1}
\alias{simu_data_M1}

\title{
Data for Model 1
}
\description{
Data simulated under Model 1.
}

\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}
\seealso{
\link{SimulateData}.
}

