\name{simu_data_M2}
\alias{simu_data_M2}
\title{
Data for Model 2
}
\description{
Data simulated under Model 2.
}

\details{
The columns of the data matrix are the genetic variant (V node) and phenotype nodes (T nodes).
}

\value{
Matrix
}
\author{
Md Bahadur Badsha (mbbadshar@gmail.com)
}
\seealso{
\link{SimulateData}.
}
