\name{simu_data_M3}
\alias{simu_data_M3}
\title{
Data for Model 3
}
\description{
Data simulated under Model 3.
}
\details{
The columns of the data matrix are the genetic variant (with three genotypes) and phenotype nodes.
}
\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}
\seealso{
\link{SimulateData}.
}
