\name{simu_data_multiparent}
\alias{simu_data_multiparent}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Data for the multiple-parent model
}
\description{
Data simulated under the multiple-parent model, where a phenotype node has multiple parent nodes.
}

\details{
The columns of the data matrix are the genetic variant (V node) and phenotype nodes (T nodes).
}

\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\seealso{
\link{SimulateData}.
}

