\name{simu_data_starshaped }
\alias{simu_data_starshaped}
\title{
Data for the star model
}
\description{
Data simulated under the star model, where one gene has more than two children.
}

\details{
The columns of the data matrix are the genetic variant (V node) and phenotype nodes (T nodes).
}
\value{
Matrix
}

\author{
Md Bahadur Badsha (mdbadsha@uidaho.edu)
}

\seealso{
\link{SimulateData}.
}
